Simplify dependencies in concurrent/synchronization/volatile
