+++ lib/concurrent-ruby/concurrent/synchronization.rb
@@ -1,11 +1,5 @@
-require 'concurrent/utility/native_extension_loader' # load native parts first
-
 require 'concurrent/utility/engine'
 
-require 'concurrent/synchronization/abstract_object'
-require 'concurrent/synchronization/mri_object'
-require 'concurrent/synchronization/jruby_object'
-require 'concurrent/synchronization/truffleruby_object'
 require 'concurrent/synchronization/object'
 require 'concurrent/synchronization/volatile'
