+++ lib/concurrent-ruby/concurrent/synchronization/full_memory_barrier.rb
@@ -1,3 +1,5 @@
+require 'concurrent/utility/native_extension_loader' # load native parts first
+
 module Concurrent
   module Synchronization
     case
