+++ lib/concurrent-ruby/concurrent/synchronization/object.rb
@@ -1,3 +1,5 @@
+require 'concurrent/utility/native_extension_loader' # load native parts first
+
 require 'concurrent/synchronization/safe_initialization'
 require 'concurrent/synchronization/volatile'
 require 'concurrent/atomic/atomic_reference'
